+++ prisma/migrations/20240806135915_change_event_id_on_coins_table_to_allow_nullable/migration.sql
@@ -1,0 +1,8 @@
+-- DropForeignKey
+ALTER TABLE "Coins" DROP CONSTRAINT "Coins_event_id_fkey";
+
+-- AlterTable
+ALTER TABLE "Coins" ALTER COLUMN "event_id" DROP NOT NULL;
+
+-- AddForeignKey
+ALTER TABLE "Coins" ADD CONSTRAINT "Coins_event_id_fkey" FOREIGN KEY ("event_id") REFERENCES "Events"("id") ON DELETE SET NULL ON UPDATE CASCADE;