+++ src/db/queries/index.ts
@@ -1,5 +1,0 @@
-export * from './achievements';
-export * from './coin';
-export * from './event';
-export * from './user';
-export * from './voice-level';