+++ src/db/schema/index.ts
@@ -1,5 +1,0 @@
-export * from './achievements-table';
-export * from './coin-table';
-export * from './event-table';
-export * from './user-table';
-export * from './voice-table';